scripts/requirements.txt
requests>=2.31
pyyaml>=6.0
anthropic>=0.25
# scholarly>=1.7  # optional: uncomment for Google Scholar support
